requirements.txt
fastapi
uvicorn
# Core & Tools
requests
python-dotenv

# Development (Hephaestus)
pytest

# QA Automation (Artemis)
robotframework
robotframework-requests
robotframework-seleniumlibrary